Clear screen during re-layouts in example
Doing re-layout for changes in window size may leave coordinates that are now not painted over by any active components.  (I experienced this when updating the example to have the `ls` object take up 2 grid widths and be offset 1, for example.)  It's probably a good idea to pair big re-layout&re-draw operations with a `Clear` call.