fix segfault
I ran into an error that if the `line` has only one data point, looking at `line[1]` will cause an index out of bounds error.

I suspect you wanted to look at the previous value, which would be `line[i-1]` instead of `line[1]` correct? I added a simple check not to try and calculate the value if we are on the first item. 

Please feel free to reject this PR but it fixed a crash in my program and did not alter the output in any noticeable way so I'm gonna use it regardless.